+++ src/main/java/com/mazebert/simulation/commands/InitGameCommand.java
@@ -1,5 +1,5 @@
 package com.mazebert.simulation.commands;
 
 public strictfp class InitGameCommand extends Command {
-    public long randomSeed;
+    public int randomSeed;
 }
